_backup_20251223_225820/requirements.txt
# Core LLM + orchestration
crewai
crewai-tools

# (Optional) LangChain helpers – keep for RAG and utilities
#langchain
#langchain-community
#langchain-openai

# Utilities
python-dotenv
python-docx
pypdf
requests
pydantic
tqdm
docxtpl


# Vector DB – friendly on Windows
chromadb

# Create and share custom web apps
streamlit

